+++ examples/grammars/css/css.js
@@ -157,7 +157,7 @@
     // ----------------- parser -----------------
 
     function CssParser(input) {
-        Parser.call(this, input, cssTokens, {maxLookahead: 3});
+        Parser.call(this, input, cssTokens);
         var $ = this;
 
         this.stylesheet = this.RULE('stylesheet', function() {